enable npm test
