+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "./node_modules/jasmine/bin/jasmine.js",
+    "test": "ts-node node_modules/jasmine/bin/jasmine",
     "lint": "eslint source/javascripts/*.js"
   },
   "repository": {
@@ -19,6 +19,7 @@
   "homepage": "https://github.com/codeforjapan/mapprint#readme",
   "devDependencies": {
     "@mapbox/togeojson": "^0.16.0",
+    "@types/jasmine": "^3.3.13",
     "brfs": "^2.0.2",
     "browserify": "^16.2.3",
     "eslint": "^5.16.0",
@@ -28,6 +29,7 @@
     "gulp-sass": "^4.0.2",
     "jasmine": "^3.4.0",
     "node-neat": "^2.0.0-beta.0",
+    "ts-node": "^8.3.0",
     "tsify": "^4.0.1",
     "typescript": "^3.5.2",
     "vinyl-source-stream": "^2.0.0"
