+++ spec/support/jasmine.json
@@ -1,7 +1,7 @@
 {
   "spec_dir": "spec",
   "spec_files": [
-    "**/*[sS]pec.js"
+    "**/*[sS]pec.ts"
   ],
   "helpers": [
     "helpers/**/*.js"
