+++ tsconfig.json
@@ -15,7 +15,8 @@
     /* Additional Checks */
     "noUnusedLocals": false,                /* Report errors on unused locals. */
     "noUnusedParameters": false,            /* Report errors on unused parameters. */
-    "noImplicitReturns": false             /* Report error when not all code paths in function return a value. */
+    "noImplicitReturns": false,             /* Report error when not all code paths in function return a value. */
+    "sourceMap": true
   },
   "include": [
       "./source/javascripts/**/*"
